requirements.txt
pandas-datareader==0.10.0
plotly==5.18.0
streamlit==1.29.0
PyYAML==6.0.1
